change(go): update go files
