+++ go/go.mod
@@ -1,20 +1,18 @@
 module mercari-build-training
 
-go 1.23.4
+go 1.24
+
+tool go.uber.org/mock/mockgen
 
 require (
-	github.com/labstack/echo/v4 v4.13.3
-	github.com/labstack/gommon v0.4.2
+	github.com/google/go-cmp v0.7.0
+	github.com/mattn/go-sqlite3 v1.14.24
+	go.uber.org/mock v0.5.0
 )
 
 require (
-	github.com/mattn/go-colorable v0.1.14 // indirect
-	github.com/mattn/go-isatty v0.0.20 // indirect
-	github.com/valyala/bytebufferpool v1.0.0 // indirect
-	github.com/valyala/fasttemplate v1.2.2 // indirect
-	golang.org/x/crypto v0.32.0 // indirect
-	golang.org/x/net v0.34.0 // indirect
-	golang.org/x/sys v0.29.0 // indirect
-	golang.org/x/text v0.21.0 // indirect
-	golang.org/x/time v0.8.0 // indirect
+	github.com/stretchr/testify v1.10.0 // indirect
+	golang.org/x/mod v0.18.0 // indirect
+	golang.org/x/sync v0.10.0 // indirect
+	golang.org/x/tools v0.22.0 // indirect
 )
