--- go/v2/go.mod
@@ -1,17 +1,0 @@
-module v2
-
-go 1.24
-
-tool go.uber.org/mock/mockgen
-
-require (
-	github.com/google/go-cmp v0.7.0
-	github.com/mattn/go-sqlite3 v1.14.24
-)
-
-require (
-	go.uber.org/mock v0.5.0 // indirect
-	golang.org/x/mod v0.18.0 // indirect
-	golang.org/x/sync v0.7.0 // indirect
-	golang.org/x/tools v0.22.0 // indirect
-)
